[Update]: Start implementing UI. [ArtyomRusak]

+++ AR.EPAM/AR.EPAM.Core/Entities/Auction/Comment.cs
@@ -14,6 +14,6 @@ namespace AR.EPAM.Core.Entities.Auction
         public virtual User User { get; set; }
         public int UserId { get; set; }
         public virtual Lot Lot { get; set; }
-        public int LotId { get; set; }
+        public long LotId { get; set; }
     }
 }
